players_list for easy access
When 'enter_players' is called, it will return a list of players who added their names. storing this in 'players_list' gives us easy access, and will print the players in a list. we will use this later to manipulate the list, and give players turns to guess higher lower.