more clearly defined higher_lower function
Using a try except block with a continue statement has helped clear up a lot about inputs being validated. This helps the program not run infinite loops as well, as the while true block is clearly defined. I used to think it was lazy programming, but as long as their is a clear exit/break, it is fine.